4 features extracted
On the basis of an old version without "last mention" feature. They are presented in the following order:
"length": length of the non-finite word lemma
"featurecompposition": a categorial feature. "NA" when there is no noun or adj around the non-finite verb, "before" when there is a noun of adj before the non-finite verb, "after" when it occurs after the verb
"featureargstr": a categorical feature. "No complement" when the non-finite verb doesn't have a complement, "NP" when it takes an Noun Phrase complement, "PP" when it takes a prepositional phrase as complement (or adjunct, I'm not distinguishing these two)
"tense": a categorical feature. "unknown" when I can't find tense (need further check). The possible values are "Present", "Past", "Future" and "Conditional" (for "would").